feat: build useHook for modal
