+++ constants/screenSize.ts
@@ -1,0 +1,1 @@
+export const tabletWidth = 834;
