+++ pages/index.tsx
@@ -3,10 +3,11 @@ import WhyChooseUs from "../components/WhyChooseUs";
 import Intro from "../components/Intro";
 import Template from "../components/Template";
 import NavBar from "../components/navbar";
-import useModal from "../components/navbar/modal";
+import useNavModal from "../components/navbar/useNavModal";
+import NavModal from "../components/navbar/NavModal";
 
 export default function Home() {
-  const { open: openMenuModal, Modal } = useModal();
+  const { open: openMenuModal, close: closeMenuModal, state: navModalState } = useNavModal();
   return (
     <>
       <NavBar open={openMenuModal} />
@@ -14,7 +15,7 @@ export default function Home() {
       <WhyChooseUs />
       <Intro />
       <Template />
-      <Modal />
+      {navModalState && <NavModal close={closeMenuModal} />}
     </>
   );
 }
